+++ CHANGELOG.md
@@ -6,11 +6,13 @@ All notable changes to the JavaDoc Generator IntelliJ Plugin will be documented 
 
 ### Added
 
-- JavaDoc Generation: This plugin utilizes the capabilities of ChatGPT AI to automatically generate JavaDoc documentation for a provided input method.
+- JavaDoc Generation: This plugin utilizes the capabilities of ChatGPT AI to automatically generate JavaDoc
+  documentation for a provided input method.
 
 ## [0.0.1] - Not Yet Published
 
 ### Features
 
-- JavaDoc Generation: This plugin leverages the capabilities of ChatGPT AI to produce JavaDoc documentation for a provided input method.
+- JavaDoc Generation: This plugin leverages the capabilities of ChatGPT AI to produce JavaDoc documentation for a
+  provided input method.